+++ backend/src/main/java/project/backend/domain/chat/chatmessage/dto/event/EventMessageResponse.java
@@ -1,4 +1,4 @@
-package project.backend.domain.chat.chatroom.dto.event;
+package project.backend.domain.chat.chatmessage.dto.event;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
