+++ backend/src/main/java/project/backend/global/exception/GlobalExceptionHandler.java
@@ -35,7 +35,6 @@ public class GlobalExceptionHandler {
 	public ResponseEntity<ErrorResponse> handleAuthException(AuthException ex) {
 
 		Map<String, Object> details = new HashMap<>();
-		details.put("roomId", ex.getRoomId());
 		details.put("error", ex.getErrorCode());
 
 		ErrorResponse response = ErrorResponse.toResponse(ex.getErrorCode(), details);
